Added a new field to Library for showing Last Scan time for a library.
Manage library page now shows last scan for each library. Tweaked the websocket code to make scan progress a bit more reliable.

+++ API/SignalR/MessageFactory.cs
@@ -59,7 +59,7 @@ namespace API.SignalR
             };
         }
 
-        public static SignalRMessage ScanLibraryProgressEvent(int libraryId, int progress, string seriesName)
+        public static SignalRMessage ScanLibraryProgressEvent(int libraryId, float progress, string seriesName)
         {
             return new SignalRMessage()
             {
